# swagger.yaml
openapi: 3.1.0
info:
  title: Умное БТИ
  description: MVP платформы для клиентов, исполнителей и администраторов
  version: 0.1.0
paths:
  /api/v1/auth/register/client:
    post:
      tags:
      - Auth
      summary: Регистрация клиента
      operationId: register_client_api_v1_auth_register_client_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterClientRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/client/orders:
    get:
      tags:
      - Client
      summary: List Client Orders
      operationId: list_client_orders_api_v1_client_orders_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
                title: Response List Client Orders Api V1 Client Orders Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Client
      summary: Create Order
      operationId: create_order_api_v1_client_orders_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/client/orders/{order_id}:
    get:
      tags:
      - Client
      summary: Get Order
      operationId: get_order_api_v1_client_orders__order_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Client
      summary: Update Order
      operationId: update_order_api_v1_client_orders__order_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/client/orders/{order_id}/files:
    post:
      tags:
      - Client
      summary: Upload File
      operationId: upload_file_api_v1_client_orders__order_id__files_post
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_api_v1_client_orders__order_id__files_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderFile'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Client
      summary: Get Files
      operationId: get_files_api_v1_client_orders__order_id__files_get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderFile'
                title: Response Get Files Api V1 Client Orders  Order Id  Files Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/client/orders/{order_id}/plan:
    get:
      tags:
      - Client
      summary: Get Plan Versions
      operationId: get_plan_versions_api_v1_client_orders__order_id__plan_get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: version
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPlanVersion'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/client/orders/{order_id}/plan/changes:
    post:
      tags:
      - Client
      summary: Add Plan Change
      operationId: add_plan_change_api_v1_client_orders__order_id__plan_changes_post
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavePlanChangesRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPlanVersion'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/client/orders/{order_id}/status-history:
    get:
      tags:
      - Client
      summary: Get Status History
      operationId: get_status_history_api_v1_client_orders__order_id__status_history_get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderStatusHistoryItem'
                title: Response Get Status History Api V1 Client Orders  Order Id  Status
                  History Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/client/orders/{order_id}/ai/analyze:
    post:
      tags:
      - Client
      summary: Trigger Ai Analyze
      operationId: trigger_ai_analyze_api_v1_client_orders__order_id__ai_analyze_post
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiAnalysis'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/client/orders/{order_id}/files/{file_id}:
    get:
      tags:
      - Client
      summary: Download File
      operationId: download_file_api_v1_client_orders__order_id__files__file_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/client/orders/{order_id}/ai/analysis:
    get:
      tags:
      - Client
      summary: Get Ai Analysis
      operationId: get_ai_analysis_api_v1_client_orders__order_id__ai_analysis_get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiAnalysis'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/executor/orders:
    get:
      tags:
      - Executor
      summary: List Executor Orders
      operationId: list_executor_orders_api_v1_executor_orders_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: department_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Department Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExecutorOrderListItem'
                title: Response List Executor Orders Api V1 Executor Orders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/executor/orders/{order_id}:
    get:
      tags:
      - Executor
      summary: Get Order
      operationId: get_order_api_v1_executor_orders__order_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutorOrderDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/orders:
    get:
      tags:
      - Admin
      summary: Список заказов (админ)
      operationId: list_orders_api_v1_admin_orders_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
                title: Response List Orders Api V1 Admin Orders Get
      security:
      - HTTPBearer: []
  /api/v1/admin/orders/{order_id}:
    get:
      tags:
      - Admin
      summary: Детали заказа (админ)
      operationId: get_order_api_v1_admin_orders__order_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Admin
      summary: Обновление заказа (админ)
      operationId: update_order_api_v1_admin_orders__order_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateOrderRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AdminUpdateOrderRequest:
      properties:
        status:
          anyOf:
          - $ref: '#/components/schemas/OrderStatus'
          - type: 'null'
        currentDepartmentCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Currentdepartmentcode
        estimatedPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Estimatedprice
        totalPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Totalprice
      type: object
      title: AdminUpdateOrderRequest
    AiAnalysis:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        orderId:
          type: string
          format: uuid
          title: Orderid
        decisionStatus:
          type: string
          title: Decisionstatus
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        risks:
          anyOf:
          - items:
              $ref: '#/components/schemas/AiRisk'
            type: array
          - type: 'null'
          title: Risks
        legalWarnings:
          anyOf:
          - type: string
          - type: 'null'
          title: Legalwarnings
        financialWarnings:
          anyOf:
          - type: string
          - type: 'null'
          title: Financialwarnings
        rawResponse:
          anyOf:
          - type: object
          - type: 'null'
          title: Rawresponse
      type: object
      required:
      - id
      - orderId
      - decisionStatus
      title: AiAnalysis
    AiRisk:
      properties:
        type:
          type: string
          title: Type
        description:
          type: string
          title: Description
        severity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Severity
        zone:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone
      type: object
      required:
      - type
      - description
      title: AiRisk
...
    Body_upload_file_api_v1_client_orders__order_id__files_post:
      properties:
        upload:
          type: string
          format: binary
          title: Upload
      type: object
      required:
      - upload
      title: Body_upload_file_api_v1_client_orders__order_id__files_post
...
    CreateOrderRequest:
      properties:
        serviceCode:
          type: integer
          title: Servicecode
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        districtCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Districtcode
        houseTypeCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Housetypecode
        calculatorInput:
          anyOf:
          - type: object
          - type: 'null'
          title: Calculatorinput
      type: object
      required:
      - serviceCode
      - title
      title: CreateOrderRequest
...
    ExecutorOrderDetails:
      properties:
        order:
          $ref: '#/components/schemas/Order'
        files:
          anyOf:
          - items:
              $ref: '#/components/schemas/OrderFile'
            type: array
          - type: 'null'
          title: Files
        planOriginal:
          anyOf:
          - $ref: '#/components/schemas/OrderPlanVersion'
          - type: 'null'
        planModified:
          anyOf:
          - $ref: '#/components/schemas/OrderPlanVersion'
          - type: 'null'
        statusHistory:
          items:
            $ref: '#/components/schemas/OrderStatusHistoryItem'
          type: array
          title: Statushistory
        client:
          anyOf:
          - $ref: '#/components/schemas/User'
          - type: 'null'
        executorAssignment:
          anyOf:
          - type: object
          - type: 'null'
          title: Executorassignment
      type: object
      required:
      - order
      title: ExecutorOrderDetails
    ExecutorOrderListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        status:
          type: string
          title: Status
        serviceTitle:
          type: string
          title: Servicetitle
        totalPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Totalprice
        createdAt:
          type: string
          format: date-time
          title: Createdat
        complexity:
          anyOf:
          - type: string
          - type: 'null'
          title: Complexity
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        departmentCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Departmentcode
      type: object
      required:
      - id
      - status
      - serviceTitle
      - createdAt
      title: ExecutorOrderListItem
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Order:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        clientId:
          type: string
          format: uuid
          title: Clientid
        serviceCode:
          type: integer
          title: Servicecode
        status:
          $ref: '#/components/schemas/OrderStatus'
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        districtCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Districtcode
        houseTypeCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Housetypecode
        complexity:
          anyOf:
          - type: string
          - type: 'null'
          title: Complexity
        calculatorInput:
          anyOf:
          - type: object
          - type: 'null'
          title: Calculatorinput
        estimatedPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Estimatedprice
        totalPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Totalprice
        currentDepartmentCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Currentdepartmentcode
        aiDecisionStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Aidecisionstatus
        aiDecisionSummary:
          anyOf:
          - type: string
          - type: 'null'
          title: Aidecisionsummary
        plannedVisitAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Plannedvisitat
        completedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completedat
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updatedat
      type: object
      required:
      - id
      - clientId
      - serviceCode
      - status
      - title
      - createdAt
      title: Order
...
    OrderFile:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        orderId:
          type: string
          format: uuid
          title: Orderid
        senderId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Senderid
        filename:
          type: string
          title: Filename
        path:
          type: string
          title: Path
      type: object
      required:
      - id
      - orderId
      - filename
      - path
      title: OrderFile
    OrderPlanVersion:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        orderId:
          type: string
          format: uuid
          title: Orderid
        versionType:
          type: string
          title: Versiontype
        plan:
          $ref: '#/components/schemas/Plan-Output'
      type: object
      required:
      - id
      - orderId
      - versionType
      - plan
      title: OrderPlanVersion
    OrderStatus:
      type: string
      enum:
      - DRAFT
      - SUBMITTED
      - EXECUTOR_ASSIGNED
      - VISIT_SCHEDULED
      - DOCUMENTS_IN_PROGRESS
      - COMPLETED
      - CANCELLED
      - REJECTED
      title: OrderStatus
    OrderStatusHistoryItem:
      properties:
        oldStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Oldstatus
        status:
          type: string
          title: Status
        changedByUserId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Changedbyuserid
        changedAt:
          type: string
          format: date-time
          title: Changedat
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
      type: object
      required:
      - status
      - changedAt
      title: OrderStatusHistoryItem
...
    Plan-Output:
      properties:
        meta:
          $ref: '#/components/schemas/PlanMeta'
        elements:
          items:
            anyOf:
            - $ref: '#/components/schemas/WallElement'
            - $ref: '#/components/schemas/ZoneElement'
            - $ref: '#/components/schemas/DoorElement'
            - $ref: '#/components/schemas/WindowElement'
            - $ref: '#/components/schemas/LabelElement'
          type: array
          title: Elements
      type: object
      required:
      - meta
      - elements
      title: Plan
...
    RegisterClientRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 6
          title: Password
        fullName:
          type: string
          title: Fullname
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        is_admin:
          type: boolean
          title: Is Admin
          default: false
      type: object
      required:
      - email
      - password
      - fullName
      title: RegisterClientRequest
...
    SavePlanChangesRequest:
      properties:
        versionType:
          type: string
          title: Versiontype
        plan:
          $ref: '#/components/schemas/Plan-Input'
      type: object
      required:
      - versionType
      - plan
      title: SavePlanChangesRequest
...
    UpdateOrderRequest:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        districtCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Districtcode
        houseTypeCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Housetypecode
        calculatorInput:
          anyOf:
          - type: object
          - type: 'null'
          title: Calculatorinput
      type: object
      title: UpdateOrderRequest
...
    User:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        fullName:
          type: string
          title: Fullname
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        isAdmin:
          type: boolean
          title: Isadmin
      type: object
      required:
      - id
      - email
      - fullName
      - isAdmin
      title: User
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer